pf/requirements.txt
Django==2.0.2
django-filter==1.1.0
django-widget-tweaks==1.4.1